Restructure for pkg.go.dev compatibility: move build pipeline to build folder, move server.go to root
